Move documents an index  tests to their own file
